requirements.txt
# Web Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0

# Machine Learning & Data Science
numpy==2.1.3
scikit-learn==1.6.0
scipy==1.15.2
tensorflow==2.18.0
keras==3.8.0
joblib==1.4.2

# Computer Vision
opencv-python==4.10.0.84
pillow==11.0.0

# Audio Processing
pyaudio==0.2.14
librosa==0.10.2.post1
soundfile==0.12.1
speechrecognition==3.12.0

# Natural Language Processing
nltk==3.9.1

# Document Processing (OCR)
pytesseract==0.3.13
pdf2image==1.17.0
PyMuPDF==1.25.2

# Google AI
google-generativeai==0.8.4
google-auth-httplib2==0.2.0

# Utilities
python-dotenv==1.0.1
matplotlib==3.10.0

# Windows-specific dependencies (for pywin32)
pywin32==308

# Other utilities
decorator==5.1.1
jsonpatch==1.33
threadpoolctl==3.5.0